+++ docs/release-notes/release_notes.md
@@ -12,6 +12,9 @@
 
 ### ✨ New Functionality
 
+- [Orchestration] [Add Spring AI Chat Memory support](https://github.com/SAP/ai-sdk-java/tree/main/docs/guides/SPRING_AI_INTEGRATION.md#chat-memory)
+- [Orchestration] [Prompt templates can be consumed from registry.](https://github.com/SAP/ai-sdk-java/tree/main/docs/guides/ORCHESTRATION_CHAT_COMPLETION.md#Chat-completion-with-Templates)
+- [Orchestration] [Grounding via *help.sap.com* is enabled.](https://github.com/SAP/ai-sdk-java/tree/main/docs/guides/ORCHESTRATION_CHAT_COMPLETION.md#grounding)
 - [OpenAI] [Spring AI integration for embedding calls.](../guides/SPRING_AI_INTEGRATION.md#embedding)
 
 ### 📈 Improvements